py/requirements/gpu.txt
llama_cpp_python==0.3.14